libeos-payg-codes: Initial implementation of code generation library
This is an internal helper library which implements generation and
validation of PAYG codes, according to the basic phase 1 design. It
includes unit tests to test basic functionality of the library.

No Debian packaging is included, because the library is so small that it
probably makes sense to just statically link it into whatever uses it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21616