eos-payg-generate: Add utility to generate PAYG codes
This is a simple utility which uses the code from libeos-payg-codes to
generate codes. It’s intended to be used by Endless employees to
generate sheets of codes to send to distributors of computers, or
individuals participating in the PAYG trial.

It is not intended for use outside Endless, and hence is packaged
separately from the daemon.

Use it like:
  eos-payg-generate /path/to/key 14d
to generate 256 14-day codes using the given key.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21617