eos-paygd: Use StateDirectory= to create /var/lib/eos-payg
Rather than punching a hole in the sandbox for an existing
/var/lib/eos-payg directory (with no way of creating the directory), get
systemd to create it for us, and automatically punch a hole in the
sandbox for it.

This fixes starting eos-paygd.service on a system where
/var/lib/eos-payg doesn’t already exist. It also ensures the directory
has the right ownership and permissions.

The StateDirectory= key was introduced in systemd 235, so this bumps our
systemd dependency.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21835