libeos-payg: Add rate limiting to epg_manager_add_code()
This is implemented using a simple FIFO array of the timestamps of the
most recent calls to epg_manager_add_code(). The rate limits are set to
10 attempts every 30 minutes, and these values are important for the
security guarantees (what little they are) of the code scheme used in
libeos-payg-codes: they ensure that any brute force attack trying to
verify codes will likely take longer than 1 week to find a valid code.
The idea being that if a code’s period is typically a week, it’s
impossible to do enough brute force on a single machine to get a valid
code for the week.

Currently, the rate limiting parameters are hard-coded, and are not
configurable.

A following commit will hook this up to D-Bus.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21766