debian: Lower systemd requirement to 232 for eos3.3
This requires reworking how the state directory is created, since the
StateDirectory= key was only added in systemd 235. Drop that, and use
tmpfiles.d instead to create the directory.

This patch should not be applied to master, as we can depend on systemd
235 there, and things are a bit cleaner.

It is possible that this could cause eos-paygd to fail to start up iff
systemd-tmpfiles does not run beforehand. Since systemd-tmpfiles runs on
every startup, I believe this can only happen if eos-paygd is upgraded
(from before this commit, to after it) after booting; or if
/var/lib/eos-payg is explicitly deleted. Neither of those situations
should affect normal users.

Signed-off-by: Philip Withnall <withnall@endlessm.com>
https://phabricator.endlessm.com/T21730