libeos-payg: Poll for the expiry time to avoid issues with suspend
Rather than setting a single long timeout period and then expiring when
that is reached, poll every 60s to see if the expiry time has been
reached. This avoids the timeout going out of sync with the wall clock
when resuming from suspend, and also works around issues with the system
clock and timezone changing.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22074