Add timeout source using CLOCK_BOOTTIME
GLib's own g_timeout_add() functions use the main loop's own idea of
time, which is defined to be CLOCK_MONOTONIC. CLOCK_MONOTONIC does not
make progress while the system is suspended. This means that
g_timeout_add (n); will call its callback no sooner than n wall-clock
milliseconds from now (modulo adjtime(3)) but may call it much later if
the computer is suspended in the interim.

CLOCK_BOOTTIME is defined to be equal to CLOCK_MONOTONIC, except that it
also includes any time that the system is suspended. So if the system
suspends and resumes before the deadline, the callback will not be
delayed; if the deadline is reached while the system is suspended, the
callback will be called as soon as the system resumes.

https://phabricator.endlessm.com/T22073