libeos-payg: Drop a FIXME comment
As wjt has pointed out https://www.evanjones.ca/linear-search.html, we
can use the plots there to argue that linear search is fine for lists up
to 256 elements; it’s the fastest search up to around 100 elements, and
competitive for array sizes a bit larger than that.

Also, this isn’t performance critical, and hasn’t been found to be slow
in practice, so the FIXME seems unwarranted in several respects.

Signed-off-by: Philip Withnall <withnall@endlessm.com>