manager: load key directly
This key is specific to this system of PAYG codes. We preserve the same
requirement that the key must be present if PAYG is enabled, but now
always try to load the key even if PAYG is disabled. This brings us
closer to a model where the presence of the key implies that PAYG is
enabled.

https://phabricator.endlessm.com/T23606