manager: set :enabled to FALSE if key is missing
The lack of a key is sufficient reason to believe that (this flavour of)
PAYG is not in use on this system. For backwards-compatibility with any
existing deployments which have a key file but where the config file has
been edited to disable PAYG, the :enabled property can still be set at
construct-time.

https://phabricator.endlessm.com/T23606