Add libpeas-based loader for external EpgProviders
Plugins will link against libeos-payg, which as a result is now
a shared library, installed to a private directory. The daemon and
tests have their rpath set accordingly when installed.

To make developing a branch of eos-payg in parallel with an external
provider easier, the $EPG_MODULE_DIR environment variable can be used to
override the default directory.

This will be wired up to EpgService in a future commit.

https://phabricator.endlessm.com/T23606