meson: allow overriding sysusersdir/systemdsystemunitdir
When fetched from systemd's .pc file, these are in system-wide
locations. This prevents `ninja install`ing the project to a --prefix.
(In particular, GNOME Builder's "Run" button installs the project to a
local prefix, but this triggers a Polkit dialog asking to install these
files system-wide.)

With this change, it's possible to pass additional arguments to `meson
[configure]` to override these locations.

https://phabricator.endlessm.com/T23606