meson: default to default_library=static
This is already forced in debian/rules to force libgsystemservice to be
a static library (so we don't have to install it). I got confused during
development because libgsystemservice.so was being installed (and
required by eos-paygd1); this was because I had not set this flag, and
so Meson defaulted to default_library=shared.

https://phabricator.endlessm.com/T23606