provider: replace save_state with shutdown
One out-of-tree provider depends on a library which uses global state.
This means that it is not possible to have multiple instances at once,
and we need to run its teardown functions exactly once.

As it happens, in the real application save_state_async() is called
exactly once, at shutdown; rename it accordingly, and make EpgManager
use an internal method to periodically save state.

https://phabricator.endlessm.com/T23606