provider: add :code-format property
The unlock UI pre-validates codes before calling AddCode(). It currently
hard-codes the 8-digit requirement, but alternative providers have
different requirements. Add a property to allow them to be discovered,
and expose it on the bus.

This is a moderate amount of boilerplate to add a single
constant string property to the D-Bus service, huh.

https://phabricator.endlessm.com/T23606