Warn on discarding const qualifier
The use of -Wno-discarded-qualifiers prevented the compiler catching a
case where I assigned the value returned by g_getenv() and a string
literal to a 'char *' variable. As it happens, I'd also forgotten to
free it, so in fact this was benign (but an explicit value would have
been leaked).

Re-enable this warning.  There was only one place in the codebase where
this warning triggered:

    ../libeos-payg/manager.c:1018:35: error: passing argument 1 of
        ‘memcpy’ discards ‘const’ qualifier from pointer target type
        [-Werror=discarded-qualifiers]
           memcpy (expiry_time_secs.u8, data, sizeof (expiry_time_secs.u8));
                   ~~~~~~~~~~~~~~~~^~~

I think this warning is valid; this field should not be declared const.
Fix that, too.