Update libgsystemservice submodule to install nothing
I realised it would be a reasonably small change to teach
libgsystemservice to behave differently when used as a Meson subproject,
as we do here.

Since gtk-doc documentation is built at install time with Meson, this
allows us to drop the gtk-doc-tools dependency. We can also drop the
not-installed file.

https://gitlab.gnome.org/pwithnall/libgsystemservice/merge_requests/13