service: exit after 30 seconds if not enabled
Previously, the shell would check the config file and only ask the
daemon if PAYG is enabled (causing it to auto-start) if the config file
says PAYG is enabled, to save running the daemon at all on a non-PAYG
system. Since we now allow plugins to report that PAYG is enabled
regardless of what the config file says, the shell will need to change
to always query the daemon.

To avoid continuously running the (idle) daemon on non-PAYG systems,
quit after 30 seconds' inactivity in that case.

https://phabricator.endlessm.com/T23606