provider: make :code-format be a regular expression
An out-of-tree provider has surprisingly complicated rules, with two
possible lengths and optional hyphens. Rather than inventing our own DSL
to encode this, it seems more straightforward to use an existing DSL for
text patterns. (The existence of G_REGEX_MATCH_PARTIAL makes this a
suitable representation for verifying the code character-by-character as
it is entered.)

https://phabricator.endlessm.com/T23965