eos-payg.conf: set Enabled=true by default
Since 0856b8db517691ac7e0a6361078e3b817b1976a0, the default provider is
enabled if and only if the key exists AND Enabled=true in the config
file. Endless images do not (by default) include a PAYG key. It follows
that we can set Enabled=true in the default config file, meaning that
provisioning a system for Endless PAYG only involves creating the key
file.

https://phabricator.endlessm.com/T23606