libeos-payg: Add a Clock GInterface
This commit adds an interface that provides two ways to track the
passage of time, one that monotonically increases and one that
represents wall clock time (which might jump if the user changes it),
and a timeout GSource based on the former view of time. This interface
will have at least two implementations, a production one that will use
CLOCK_BOOTTIME and the system clock time, and an implementation for unit
tests that allows the clocks to be fast forwarded (or rewinded).

https://phabricator.endlessm.com/T24300