Add an EpgRealClock implementation
This adds an implementation for the Clock interface which is intended to
be used in production, so it uses CLOCK_BOOTTIME for the get_time()
function and the system clock for get_wallclock_time().
EpgBoottimeSource is used to provide a GSource. The reason we use
CLOCK_BOOTTIME is so that it monotonically increases even while the
system is suspended. See the commit message that introduced
EpgBoottimeSource for more information.

This will be used in a subsequent commit.

https://phabricator.endlessm.com/T24300