Add and use a clock property on EpgProvider
This commit adds a property to the EpgProvider interface called "clock"
which takes an object that implements the EpgClock interface. Also,
modify EpgManager to take a clock in its constructor which goes into
said property, and update the unit tests. A subsequent commit will make
use of the clock as an alternative to g_get_real_time(),
g_timeout_source_add(), etc.

https://phabricator.endlessm.com/T24300