manager: Use CLOCK_BOOTTIME during runtime
Currently EpgManager uses the system time (via g_get_real_time()) to
determine when the expiration time will be, and to schedule timeouts to
consume credit. This is undesirable because it is susceptible to the
user changing the clock time. So this commit changes EpgManager so that
it uses the get_time() function on the EpgClock interface (which in the
case of EpgRealClock means using CLOCK_BOOTTIME) to calculate when the
expiration will happen. This means these calculations are unaffected by
changes to the system clock. We're also using CLOCK_BOOTTIME to
create timeout events (via the source_new_seconds() interface).

Importantly CLOCK_BOOTTIME includes time that passes while the computer
is suspended, unlike CLOCK_MONOTONIC.

Since CLOCK_BOOTTIME and therefore self->expiry_time_secs are relative
to an epoch that differs across reboots, it's not suitable for storage
when we're saving state. So this commit also changes the state saving
mechanism to save the wall clock time and seconds left to expiration (in
separate files) and only keeps support for the expiry time state file
for backwards compatibility. This new scheme allows us to tell on
startup whether time went forwards or backwards while the computer was
off. If time went backwards there are two possibilities:
1. The current time is incorrect, in which case NTP should correct it
(see T24501 for discussion of consuming credit for the correction)
2. The saved time is incorrect, in which case there's no way to know how
much time passed since the last state save, and the way it's implemented
we assume no time passed. This is one of the weaknesses of the system.

Since the get_expiry_time() interface of EpgProvider exposes the expiry
time over D-Bus and the shell is the consumer of this information, the
shell will have to be changed to understand the value in terms of
CLOCK_BOOTTIME instead of the system time.

https://phabricator.endlessm.com/T24300