Remove "now" parameter from epg_provider_add_code()
Currently the epg_provider_add_code() interface has a parameter that
allows the caller to specify the current time, in order to make it
easier to write unit tests that fast forward time. But now that we have
EpgFakeClock, we can use that in the unit tests instead. This commit
removes the "now" parameter from epg_provider_add_code() and changes the
unit tests to use EpgFakeClock instead. This makes things work more
intuitively but is otherwise a minor change.

https://phabricator.endlessm.com/T24300