tests: Add new state files to tests
This informs the test infrastructure about the two new state files,
clock-time and expiry-seconds, so they are deleted when cleaning up a
temporary directory, and so we test what happens when they have invalid
values. Leave the testing in place for the "expiry-time" file because we
support it for backwards compatibility.

https://phabricator.endlessm.com/T24300