Fix tests for malformed state files
In the tests we test what happens when each state file has malformed
data, but this is broken in two cases because of recent changes:

1. For the expiry-time file, we need to return G_IO_ERROR_INVALID_DATA
rather than assuming it's being deleted because it's deprecated and
returning no error. So change the machinery around file_load_delete_cb
to make this happen.

2. For the expiry-seconds file, it's only read if clock-time exists,
since that's the only case where it has useful information. So add a
special case in the tests to create clock-time so that we can test what
happens when expiry-seconds is malformed.

https://phabricator.endlessm.com/T24300