Generate GDBusInterfaceInfo struct, if possible
gdbus-codegen from GLib 2.58.3 will support generating just the
GDBusInterfaceInfo from D-Bus introspection XML, without all the other
boilerplate for client and service implementations.  This codebase
previously contained a hand-written copy of the GDBusInterfaceInfo, with
a FIXME about generating it in future.

At some point, we will want to add more D-Bus API to this daemon. It
would be nice to be able to make use of this code-generation when that
time comes, rather than having to wait for a new GLib in the distro. So:

* Add a (not-installed) XML description of the D-Bus API
* Add a static copy of the generated code
* Add Meson logic to check if a new-enough gdbus-codegen is available
  * If not, use the checked-in copy of the generated code
  * If so:
    * Generate the code
    * Make the test suite fail if the generated code doesn't match the
      checked-in code (to try to avoid them getting out of sync)
    * Emit a warning encouraging future generations to remove the
      fallback logic and checked-in code

https://phabricator.endlessm.com/T23781