manager: save state on startup
On startup, EpgManager reads the state files and consumes credit
according to how much time passed while the computer was off. We then
need to save state again because otherwise if the computer shuts down
uncleanly we would consume credit for the same time period again. So
this commit adds a save_state_async() call in file_load_cb() after the
credit has been consumed for the difference between the current time and
the time of last state save.

https://phabricator.endlessm.com/T24300