eos-payg-ctl: Understand CLOCK_BOOTTIME properties
Now that the "ExpiryTime" and "RateLimitEndTime" properties of
EpgProvider are in terms of CLOCK_BOOTTIME instead of being UNIX
timestamps, modify eos-payg-ctl to interpret them correctly. For now
this is implemented using a helper program until we get Python 3.7 on
the stable branch, which will give us time.CLOCK_BOOTTIME (to be used
with time.clock_gettime()).

https://phabricator.endlessm.com/T24300