eos-payg-ctl: Use pure Python to get CLOCK_BOOTTIME
It turns out you can pass the value "7" directly to time.clock_gettime()
rather than waiting for Python 3.7 to define time.CLOCK_BOOTTIME as 7.
So do this rather than using a helper program written in C. This is also
needed because the C program was added to the eos-payg-data package
which is architecture independent, so an ARM binary ended up in the
Intel ostree (leading to a runtime error).

https://phabricator.endlessm.com/T24300