eos-payg-ctl: use same reference timestamps when converting to UTC
Before this change, ExpiryTime and RateLimitEndTime were converted from
CLOCK_BOOTTIME to UTC by checking the current time (on both clocks)
afresh each time. This led to output like the following, with fractional
differences between the two timestamps, even when the values are
identical (in this case, 0):

          CodeFormat: ^[0-9]{8}$
             Enabled: False
          ExpiryTime: 2019-01-09 12:46:56.446161+00:00
    RateLimitEndTime: 2019-01-09 12:46:56.446158+00:00

With this change, the output looks like this:

          CodeFormat: ^[0-9]{8}$
             Enabled: False
          ExpiryTime: 2019-01-09 12:46:56.446159+00:00
    RateLimitEndTime: 2019-01-09 12:46:56.446159+00:00

It's a pedantic change, but sometimes I just can't resist.

https://phabricator.endlessm.com/T24300