.service: remove PrivateDevices, add DevicePolicy
An out-of-tree provider needs to be able to enumerate the contents of
/dev; PrivateDevices=yes causes a minimal /dev to be mounted for the
executed process, which means the important devices are not visible to
eos-paygd.

PrivateDevices=true implies a number of other restrictions:

* @raw-io system calls are blocked
* CAP_MKNOD and CAP_SYS_RAWIO are removed from CapabilityBoundingSet
* DevicePolicy=closed is set

The first two restrictions are already by other lines in this file; add
DevicePolicy=closed. (The out-of-tree provider opens the relevant device
using UDisks, essentially circumventing this restriction, but in a
controlled way.)

https://phabricator.endlessm.com/T24134