Add and implement epg_provider_wallclock_time_changed()
When the system clock changes discontinuously, either because NTP made
an adjustment or the user did, pay-as-you-go providers need to know
about it so they can save state as soon as possible. This ensures that
the credit consumption calculation on the next boot for the time spent
off will be accurate, as long as the clock kept ticking while the
computer was off. As a second line of defense, we can also consume
credit for the delta of the clock change when it's a forward jump,
because that may mean we miscalculated the credit consumption at
startup, in case we somehow were not able to save state before shutdown.

Concretely, if it's 9 PM and I set the clock back two hours to 7 PM and
immediately shut down and leave it shut down for two hours, there are
two possibilites for the credit consumption at startup when the clock
says 9 PM:
1) If the state was not saved at shut down, it will appear that no time
passed and those two hours will be unaccounted for (unless the clock is
moved forward eventually in the OS not the BIOS by NTP or the user).
2) If the state was saved at shut down, we will correctly calculate that
two hours passed (even though the absolute numbers are wrong).

So this commit adds a method to the EpgProvider interface,
wallclock_time_changed() which informs a provider that the clock changed
discontinously, and by how much. This commit also implements that
function in EpgManager which consumes credit for positive deltas and
saves state for all nonzero deltas. A subsequent commit will add the
infrastructure for calling this function at the appropriate time.

https://phabricator.endlessm.com/T24501