service: Detect discontinous clock changes
This commit modifies EpgService so that on startup it creates a GSource
using epg_clock_jump_source_new(), and when that source fires, it
calculates the delta of the clock change and informs the loaded
EpgProvider about it. This will allow state to be saved as soon as the
clock jumps; see the discussion on the ticket for why that's necessary.

I spent a while trying to write a unit test for this, but couldn't find
a way. At the moment in the EpgService unit test we only do
epg_service_new() but don't actually get the service running on D-Bus,
so the code that initializes the GSource is never run. For various
reasons it's not as simple as just running gss_service_run().

https://phabricator.endlessm.com/T24501