eos-paygd: Don't exit after inactivity
Currently eos-paygd exits after 30 seconds of inactivity. However now
that we're starting from the initramfs we want to continue running until
shutdown to ensure that we don't run the unsigned binary on the primary
root filesystem. So this commit disables the inactivity timeout. Since
the eos-payg dracut module will only be included in PAYG images, this
won't mean that eos-paygd is always running on all Endless computers.

https://phabricator.endlessm.com/T27037