Move eos-paygd to initramfs
Currently eos-paygd lives on the main root filesystem which means the
user can modify it. This commit moves eos-paygd to the initramfs so that
is signed along with the kernel, and the signature is checked at boot
time. PAYG systems will always have Secure Boot enabled and only boot
Endless images which enforce PAYG. Therefore by starting eos-paygd in
the initramfs we are ensuring its integrity.

This move is accomplished by creating a dracut module which installs
eos-paygd-2.service (see below). The dracut module will be included in
the dracut command used to build payg images.

Also for backwards compatibility, break eos-paygd into two systemd
services: (1) eos-paygd.service which will run on Phase 2 PAYG systems
and start from the primary root filesystem, and (2) eos-paygd-2.service
which will run on Phase 4+ PAYG systems and start from the initramfs. To
make this scheme work we need to change eos-paygd-2.service.in in
several ways:
- Add DefaultDependencies=no so we can start before basic.target (see
below).
- Only run from the initrd by checking for "/etc/initrd-release".
- Only run on Phase 4 systems using ConditionKernelCommandLine=
- Use FailureAction=poweroff to cause a shutdown if it fails to start up
(we don't want emergency.target because we don't want the user to have a
root shell without eos-paygd running).
- Run before basic.target and after sysinit.target, so we
can run early in the startup sequence and reduce the chance that the
user might drop to a root shell before eos-paygd starts.
- Add Conflicts=shutdown.target so we are stopped gracefully on
shutdown, despite having DefaultDependencies=no
- Change Type=dbus to Type=notify since D-Bus isn't available in the
initramfs, and we need to be able to postpone the root pivot until we're
ready to avoid a race condition.
- Run as root rather than as a daemon user, which is necessary in order
for the process to survive systemd's killing spree that happens
immediately after the root pivot.
- Add KillMode=none and SendSIGKILL=no to avoid systemd killing us while
stopping basic.target.
- Remove the [Install] section, as the unit is installed by the dracut
module, and we don't want it to be started again from the primary
filesystem.

To avoid gss_service_run() erroring out when it sees that the service is
running as root, we also have to change EpgService to use the new
"GssService:allow-root" property, and update the libgsystemservice
submodule commit so that's available.

In eos-paygd's main() we also now call sd_notify() to inform systemd
that it's okay to proceed to the pivot to the final root. This is
required of Type=notify units and will prove even more useful later when
we need to do more in the initramfs and want to avoid a race condition.

Also, make related changes to stop creating the daemon user account,
remove the D-Bus Service config file since we don't want to be D-Bus
activatable, and change the D-Bus busconfig to let root own the name
com.endlessm.Payg1.

Also, change the Debian packaging to depend on dracut and stop
installing the D-Bus service config and system user creation config.
We don't need to have a dpkg trigger to update the initramfs, since we
will explicitly include this module when building payg images (and don't
want it in others).

Also, update relevant documentation.

And finally, drop the integration tests for eos-paygd (which basically
just involved checking the exit status). Otherwise eos-paygd would see
that it's running outside the initramfs and enter backwards compat mode,
which is not what we really want to test going forward. See
https://phabricator.endlessm.com/T27360

https://phabricator.endlessm.com/T27037