eos-paygd: Delay D-Bus connection until after root pivot
Now that we're starting eos-paygd in the initramfs, we can't immediately
connect to D-Bus as there is no daemon in the initramfs. That's fine
though, we only need D-Bus to communicate with gnome-shell, so wait for
the final root pivot and for the system bus to come up before proceeding
with main(). To ensure we don't end up in an infinite loop, time out
waiting for the root pivot after 10 minutes (even fsck on a slow hard
drive should happen in that time frame), and time out waiting for the
D-Bus daemon after another 20 minutes (allowing enough time for any slow
system services). In case of timeout, we force a poweroff to ensure the
system isn't in a state where eos-paygd can't enforce PAYG.

https://phabricator.endlessm.com/T27037