service: Exit if no PAYG providers are enabled
If no nonfree PAYG providers are enabled, EpgService falls back to using
EpgManager. And previously to the commit "eos-paygd: Don't exit after
inactivity", EpgService would exit after 30 seconds of inactivity if
EpgManager is not enabled (because the computer hasn't been provisioned
with an Endless PAYG key). However now that we want the eos-paygd
process to always be running, and we have to let the daemon continue to
start from the main root filesystem to support already-deployed
computers which can't upgrade to the sd-boot/initramfs scheme[1], we're
in the unfortunate situation of having eos-paygd always running on every
Endless computer, PAYG or not. So this commit calls gss_service_exit()
in case EpgManager isn't enabled, so that the daemon gracefully exits on
non-PAYG systems.

An alternative implementation would be to re-enable the inactivity
timeout and then call gss_service_hold() and only call
gss_service_release() on non-PAYG systems, but I like this
implementation better, because it seems more readable to not set an
inactivity timeout on a service which we don't want to timeout.

[1] https://phabricator.endlessm.com/T27524

https://phabricator.endlessm.com/T27037