eos-paygd: Ensure state files are owned by root
In released versions of EOS, eos-paygd runs as the eos-paygd user and
that user owns /var/lib/eos-payg and the state files in that directory.
Since we're removing the eos-paygd user and running the daemon as root,
we should change the files' ownership accordingly. We don't want the
state files owned by some other user in case the UID or GID of eos-paygd
is ever recycled. So this commit makes the daemon run chown after the
root filesystem pivot. We can't use an ExecStartPre= in the unit file
because that would be run in the initramfs in the case of
eos-paygd-2.service.

We need the eos-paygd user gone because we use its existence or
non-existence to determine how to run the provisioning program in the
factory test script.

https://phabricator.endlessm.com/T27524