service: Don't seg fault on shutdown w/ no provider
Before commit 97d4edf36, the provider field of an EpgService object was
always non-NULL after epg_service_startup_async() completed, since an
EpgManager was used as a fallback if other providers weren't enabled,
regardless of whether or not Endless PAYG was enabled. But after commit
97d4edf36, we may complete startup with no provider if none were
enabled. So add a check to epg_service_shutdown() so we don't try to use
a NULL provider object, which I believe is causing a seg fault.

Since provider_notify_enabled_cb() is already called by
epg_service_startup_complete() and there we exit the service if the
provider is disabled, we could remove the epg_provider_get_enabled()
check in manager_new_cb() and avoid the need for this patch. But perhaps
leaving in the early enabled check is better, so we don't do superfluous
initialization of a disabled provider (especially since the whole system
goes down if eos-paygd goes down).

https://phabricator.endlessm.com/T27037