service: Fix a g_debug missing format parameters
Fix a g_debug() message which has a %s but no argument provided for it,
which causes a seg fault when a payg image is booted with no providers
enabled.

It's frightening that any crashing bug in eos-paygd brings the whole
system down. I filed this ticket about doing an audit:
https://phabricator.endlessm.com/T27681

https://phabricator.endlessm.com/T27581