meson.build: Change warning_level to 3 and werror to true
Increase the warning level to 3. Meson decides which -W cflags that
translates to. And set werror to true, which translates to -Werror.

Also update libgsystemservice to get some fixes for warnings that are
enabled by this change.

The impetus for this is that any bugs in eos-paygd have potentially
severe impact (e.g. not enforce PAYG or crash and temporarily brick a
computer).