eos-paygd: Ping watchdog timer every 10 seconds
In case eos-paygd is running from the initramfs, open /dev/watchdog and
write a byte to it every 10 seconds. The open happens before the pivot
out of the initramfs so we can be sure it will not be interefered with.
Once the watchdog device is open it has to be pinged periodically to
prevent a shutdown, so if eos-paygd stops functioning for any reason the
computer will shutdown.

The endlessdog kernel module is only loaded for the payg initramfs, not
for non-PAYG Endless images, so that's why it makes sense to only open
the watchdog if started from the initramfs.

https://phabricator.endlessm.com/T27051