eos-paygd: Ping watchdog while waiting for system init
The watchdog timer ping happens in a callback added to the GMainContext
with g_timeout_add_seconds_full() so it only gets called if the
GMainContext is iterated. In eos-paygd's main() we have a couple while
loops which run for up to 10 or 20 minutes but don't iterate the main
loop, which means we would go that long without pinging the watchdog
timer. This commit changes both while loops to iterate the main loop so
we ping the watchdog timer every 10 seconds. I'm going to propose
increasing the watchdog timeout from 1 minute to 20 or 30, so the
frequent pinging will be less necessary, but it's still the intended
behavior, and more robust in case we decide to change either the
watchdog timeout or the timeout for these loops.

https://phabricator.endlessm.com/T27051