eos-paygd: Relax sandboxing to allow chown of state directory
Machines provisioned before 3.7.0 have their /var/lib/eos-payg directory
owned by the eos-paygd user, but eos-paygd now runs as root so we try to
chown the directory on startup. However this currently fails with the
error message:

/bin/chown: cannot read directory '/var/lib/eos-payg': Permission denied

That failure is not treated as fatal, but then when EpaProviderAngaza
tries to read its state, it hits permission denied errors again, and the
whole eos-paygd.service goes down (and PAYG goes unenforced).

The reason for the failure is the strict sandboxing by systemd we have
on eos-paygd, so loosen the sandboxing so we can chown the directory.

These changes apply to eos-paygd-2.service as well (the service run from
the initramfs) because it too chowns the state directory.

https://phabricator.endlessm.com/T27037