eos-paygd: Don't rely on systemd to create the state directory
Currently we use the StateDirectory= directive to instruct systemd to
create the directory /var/lib/eos-payg for us, and the provider
implementation expects that to exist already. However this can't
possibly work for eos-paygd-2.service, since the directory will be
created in the initramfs (or not at all), and not in the final root
filesystem. So remove the StateDirectory= and StateDirectoryMode=
options from eos-paygd-2.service.in, and have eos-paygd create the
directory itself. Since eos-paygd does this regardless of if it's in
backward compatibility mode or not, make the same changes to
eos-paygd.service.in.

https://phabricator.endlessm.com/T27037