eos-paygd: Change DevicePolicy to auto
Currently in eos-paygd{-2}.service.in we have DevicePolicy=closed, which
means with a few exceptions that eos-paygd can't access the device files
in /dev/. But we need to be able to access /dev/mmcblkxboot0, and since
we can't make an assumption about the value of x (usually 0), we can't
use DeviceAllow= to make an exception for that boot partition device. So
change DevicePolicy to auto, which means we're allowed access to all
devices as long as we don't use DeviceAllow. In the case of
eos-paygd-2.service (the initramfs service) we also need access to
/dev/watchdog. Currently payg images fail to boot successfully with the
error "eos-paygd could not open /dev/watchdog: Operation not permitted".

I tested this fix in a VM, albeit with an eos image not a payg one.

https://phabricator.endlessm.com/T27051