eos-paygd: Exit gracefully on non-PAYG systems
In case no PAYG providers are found to be enabled, EpgService takes one
of two paths: (1) in case the EOSPAYG_active EFI variable is set, the
EPG_SERVICE_ERROR_NO_PROVIDER error is returned which causes eos-paygd
to force a poweroff, since that EFI variable being set means PAYG should
be enabled, or (2) in case the EOSPAYG_active EFI variable is not set,
gss_service_exit() is called with no error set to exit from
gss_service_run(). However in this second case an error is set by
gss_service_exit() which calls g_cancellable_cancel(), so even though
the service didn't set an error there is one set for the caller of
gss_service_run(), in this case main().

So this commit updates the libgsystemservice submodule to pull in this
patch[1] so an error is no longer returned from gss_service_run(). This
means on non-PAYG systems (or PAYG systems that have been paid off),
eos-paygd.service will be considered to have succeeded by systemd.

This patch is especially important now that we're using a watchdog
timer, because we only continue pinging the watchdog indefinitely if no
errors were encountered; on error eos-paygd exits and the system shuts
down when the watchdog expires.

https://phabricator.endlessm.com/T27836